+++ package.json
@@ -4,6 +4,9 @@
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.10.6",
+    "@fortawesome/fontawesome-svg-core": "^6.3.0",
+    "@fortawesome/free-solid-svg-icons": "^6.3.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "@mantine/core": "^6.0.1",
     "@mantine/hooks": "^6.0.1",
     "@reduxjs/toolkit": "^1.9.3",
